Move test code around into HostEntry
