+++ src/hostentry.rs
@@ -105,3 +105,49 @@ impl fmt::Display for HostEntry {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+
+    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+
+    use crate::hostentry::HostEntry;
+
+    #[test]
+    fn test_ip_addr() {
+        let localhost_v4 = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
+        let localhost_v6 = IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1));
+
+        assert_eq!("127.0.0.1".parse(), Ok(localhost_v4));
+        assert_eq!("::1".parse(), Ok(localhost_v6));
+    }
+
+    #[test]
+    fn test_hostentry_has_name() {
+        let he = HostEntry {
+            ip: None,
+            name: Some(String::from("arjen.wiersma.nl")),
+            aliasses: None,
+            comment: None,
+        };
+
+        assert!(he.has_name("arjen.wiersma.nl"));
+        assert!(!he.has_name("arjen2.wiersma.nl"));
+        assert!(!he.has_name("wiersma.nl"));
+
+        let ahe = HostEntry {
+            ip: None,
+            name: Some(String::from("jelle.wiersma.nl")),
+            aliasses: Some(vec![
+                String::from("arjen.wiersma.nl"),
+                String::from("rebecca.wiersma.nl"),
+            ]),
+            comment: None,
+        };
+        assert!(ahe.has_name("arjen.wiersma.nl"));
+        assert!(ahe.has_name("rebecca.wiersma.nl"));
+        assert!(!ahe.has_name("arjen2.wiersma.nl"));
+        assert!(!ahe.has_name("wiersma.nl"));
+        assert!(ahe.has_name("jelle.wiersma.nl"));
+    }
+}
